Resume Builder/requirements.txt
streamlit==1.18.1
python-dotenv==1.0.0
google-generativeai
fpdf==1.7.2
